Minor change to support count adjacent mines and renaming

+++ minesweeper-server/src/main/java/com/minesweeper/domain/minefield/MinefieldConfig.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public abstract class MinefieldConfig {
 	private MinefieldType type;
-	private int minesCount;
+	private long minesCount;
 }
